feat: split MainActivity into multiple subclasses and make cache work again
