+++ app/src/main/java/de/uniks/codecamp/group_a/weather/viewmodel/WeatherViewModel.kt
@@ -1,17 +1,8 @@
 package de.uniks.codecamp.group_a.weather.viewmodel
 
-import androidx.compose.runtime.State
-import androidx.compose.runtime.mutableStateOf
-import android.app.Application
-import android.content.Context
-import android.location.Location
-import android.location.LocationManager
 import androidx.compose.runtime.*
-import androidx.core.content.ContextCompat
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.google.accompanist.permissions.PermissionState
-import dagger.hilt.android.internal.Contexts.getApplication
 import dagger.hilt.android.lifecycle.HiltViewModel
 import de.uniks.codecamp.group_a.weather.domain.repository.WeatherRepositoryInterface
 import de.uniks.codecamp.group_a.weather.util.Resource
@@ -19,9 +10,6 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -39,15 +27,11 @@ class WeatherViewModel @Inject constructor(
         data class ShowSnackbar(val message: String) : UIEvent()
     }
 
+
     var isRefreshing by mutableStateOf(false)
 
-    init { // Initially load weather data
-        loadAll()
-    }
-
-    fun loadAll() {
+    init {
         loadCurrentWeather()
-        loadForecast()
     }
 
     fun loadCurrentWeather() {
@@ -65,7 +49,7 @@ class WeatherViewModel @Inject constructor(
                             weatherDataItems = result.data ?: emptyList(),
                             isLoading = false
                         )
-                        _eventFlow.emit(UIEvent.ShowSnackbar("Couldn't load current weather data. Make sure to enable Location Access."))
+                        _eventFlow.emit(UIEvent.ShowSnackbar("Couldn't retrieve fresh weather data. Check your internet connection and make sure to enable Location Access."))
                     }
                     is Resource.Loading -> {
                         _weatherDataState.value = weatherDataState.value.copy(
